% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Compound}
\alias{Compound}
\alias{compound}
\alias{or}
\alias{and}
\alias{not}
\title{Build a custom compound operator.}
\usage{
compound(sep, prefix = TRUE, group = TRUE, max_length = Inf)

or(...)

and(...)

not(...)
}
\arguments{
\item{sep}{Character(1) operator name.}

\item{prefix}{Logical(1) should operator be at beginning?}

\item{group}{Logical(1) should the parts be wrapped in parens?}

\item{max_length}{Integer(1) maximum number of named arguments.}
}
\value{
Function -> Character
}
\description{
Build a custom compound operator.
}
