% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gnu_find}
\alias{gnu_find}
\title{Wrapper around GNU Find CLI utility.}
\usage{
gnu_find(search_path = getwd(), ..., find_exec = "find",
  debug = FALSE)
}
\arguments{
\item{search_path}{Character path names.}

\item{...}{Arguments to find.}

\item{find_exec}{Character(1) path to GNU Find executable.}

\item{debug}{Logical(1) whether to print generated command.}
}
\value{
Character.
}
\description{
It's designed so that arguments are just passed in
order much like the GNU Find utility itself.
}
\details{
Hopefully this will be somewhat useful to someone.

There are two special kinds of argument names that may be repeated:

- switch (sw) :: switches which don't have arguments
- compound (cmp) :: compound statements

These enable this wrapper to handle the a broader chunch of the
Find utilities functionality.

Please refer the GNU Find INFO pages for more details.
}
\examples{
# Find all non-hidden directories in $HOME.
gnu_find(search_path = "~",
         type = "d", maxdepth = 1,
         compound = not(path = "*/\\\\.*"))
}
